+++ src/main/java/de/maxhenkel/shulkerbox/Events.java
@@ -1,40 +1,28 @@
 package de.maxhenkel.shulkerbox;
 
 import de.maxhenkel.shulkerbox.gui.ShulkerboxContainer;
-import net.minecraft.block.ShulkerBoxBlock;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.ItemUseContext;
+import net.minecraft.util.ActionResultType;
 import net.minecraft.util.Hand;
 import net.minecraftforge.common.util.FakePlayer;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
-import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mod.EventBusSubscriber(modid = Main.MODID)
 public class Events {
 
-    @SubscribeEvent
-    public void onRightClick(BlockEvent.EntityPlaceEvent event) {
-        if (event.isCanceled()) {
-            return;
-        }
-
-        if (!(event.getEntity() instanceof PlayerEntity)) {
-            return;
-        }
-
-        PlayerEntity player = (PlayerEntity) event.getEntity();
+    public static void onPlaceBlock(ItemUseContext context, CallbackInfoReturnable<ActionResultType> cir) {
+        PlayerEntity player = context.getPlayer();
 
         if (player instanceof FakePlayer) {
             return;
         }
 
-        if (!(event.getPlacedBlock().getBlock() instanceof ShulkerBoxBlock)) {
-            return;
-        }
-
-        ItemStack stack = Utils.getShulkerBox(player);
+        ItemStack stack = Utils.getShulkerBox(player, context.getHand());
         if (stack == null) {
             return;
         }
@@ -42,42 +30,10 @@ public class Events {
         if (Main.SERVER_CONFIG.onlySneakPlace.get()) {
             if (!player.isSneaking()) {
                 Utils.openShulkerBox(player, stack);
-                event.setCanceled(true);
+                cir.setReturnValue(ActionResultType.SUCCESS);
             }
         }
-
     }
-
-  /*  @SubscribeEvent
-    public void onRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
-        if (event.isCanceled()) {
-            return;
-        }
-
-        PlayerEntity player = event.getPlayer();
-        Hand hand = event.getHand();
-
-        if (player instanceof FakePlayer) {
-            return;
-        }
-
-        if (hand.equals(Hand.OFF_HAND) && Utils.getShulkerBox(player, Hand.MAIN_HAND) != null) {
-            event.setCanceled(true);
-            return;
-        }
-
-        ItemStack stack = Utils.getShulkerBox(player, hand);
-        if (stack == null) {
-            return;
-        }
-
-        if (Config.SERVER.onlySneakPlace.get()) {
-            if (!player.func_225608_bj_()) {
-                Utils.openShulkerBox(player, stack);
-                event.setCanceled(true);
-            }
-        }
-    }*/
 
     @SubscribeEvent
     public void onRightClickItem(PlayerInteractEvent.RightClickItem event) {